Sorted out parent/child relationship with TimelineConfigWidget and TimelineWidget.  Timeline now shows up in TimelineTab and is updated when opened.  Added a separate data structure (ActionDataStruct) to hold info about actions added to timeline.  Should take care of a lot of infrastructure to support action loading.

+++ projects/ui/src/widgets/TimelineEntries/TimelineEntry.h
@@ -39,26 +39,28 @@ public:
   virtual QSize minimumSizeHint() const = 0;
   virtual QSize sizeHint() const = 0;
 
-  int X() const { return _x; };
-  int Y() const { return _y; };
-  int Width() const { return _width; };
-  int Height() const { return _height; };
-  TimelineEntry& X(int x)
+  double X() const { return _x; };
+  double Y() const { return _y; };
+  double Width() const { return _width; };
+  double Height() const { return _height; };
+
+  TimelineEntry& X(double x)
   {
     _x = x;
     return *this;
   };
-  TimelineEntry& Y(int y)
+  TimelineEntry& Y(double y)
   {
     _y = y;
     return *this;
   };
 
+
 protected:
-  int _x;
-  int _y;
-  int _height;
-  int _width;
+  double _x;
+  double _y;
+  double _height;
+  double _width;
 };
 }
 #endif
